Move singleton parent class to utils.py
