+++ dolphin/drivers/manager.py
@@ -11,31 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import threading
-
 from oslo_log import log
 
+from dolphin import utils
 
 LOG = log.getLogger(__name__)
-lock = threading.Lock()
 
 DRIVER_MAPPING = {
     "fake_storage": "dolphin.drivers.fake_storage.FakeStorageDriver"
 }
 
 
-class Singleton(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            with lock:
-                if cls not in cls._instances:
-                    cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-
-class DriverManager(metaclass=Singleton):
+class DriverManager(metaclass=utils.Singleton):
 
     def __init__(self):
         # The driver_factory will keep the driver instance for
